third_party/WebKit/LayoutTests/fast/css/scroll-blocks-on-parsing-expected.txt

Test the parsing and application of the scroll-blocks-on property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test case: default
PASS style.scrollBlocksOn is "none"

Test case: stylesheet-wheel
PASS style.scrollBlocksOn is "wheel-event"

Test case: explicit-none
PASS style.scrollBlocksOn is "none"

Test case: explicit-start-touch
PASS style.scrollBlocksOn is "start-touch"

Test case: explicit-wheel-event
PASS style.scrollBlocksOn is "wheel-event"

Test case: explicit-scroll-event
PASS style.scrollBlocksOn is "scroll-event"

Test case: explicit-all
PASS style.scrollBlocksOn is "start-touch wheel-event scroll-event"

Test case: explicit-touch-wheel
PASS style.scrollBlocksOn is "start-touch wheel-event"

Test case: explicit-invalid-bogus
PASS style.scrollBlocksOn is "none"

Test case: explicit-invalid-combination
PASS style.scrollBlocksOn is "none"

Test case: explicit-invalid-override
PASS style.scrollBlocksOn is "wheel-event"

Test case: explicit-invalid-duplicate
PASS style.scrollBlocksOn is "none"

Test case: explicit-invalid-bogus-list
PASS style.scrollBlocksOn is "none"

Test case: not-inherited
PASS style.scrollBlocksOn is "none"

Test case: inherit
PASS style.scrollBlocksOn is "wheel-event"

Test case: initial
PASS style.scrollBlocksOn is "none"

PASS successfullyParsed is true

TEST COMPLETE
